Declare the migration's count columns as bigint, not int
Schema verification rejected the artist group before writing anything:

    album_count (expected type: 'int', derived from input: 'bigint')
    track_count (expected type: 'int', derived from input: 'bigint')

pandas defaults integer columns to int64, which hsfs derives as 'bigint';
'int' is its name for int32. line_count on the purchases group had the same
defect and would have failed next.

Rather than fix the two columns the job happened to reach, I checked every
explicitly declared feature against the type hsfs derives from the actual
pandas dtype (int <- int8/16/32 and uint8/16, bigint <- int64 and uint32,
string <- str, timestamp <- datetime). Three columns were wrong, all the
same cause; everything else already matched. The audit now reports no
mismatches.

+++ chinook/migrate_to_feature_store.py
@@ -248,8 +248,11 @@ def main() -> None:
         online_enabled=True,
         features=[
             Feature("artist_name", type="string", online_type="varchar(200)"),
-            Feature("album_count", type="int"),
-            Feature("track_count", type="int"),
+            # bigint, not int: pandas defaults integer columns to int64, which
+            # hsfs derives as 'bigint'. Declaring 'int' (its name for int32)
+            # fails schema verification before anything is written.
+            Feature("album_count", type="bigint"),
+            Feature("track_count", type="bigint"),
             _json_feature("albums"),
             Feature("migrated_at", type="timestamp"),
         ],
@@ -268,7 +271,7 @@ def main() -> None:
             Feature("first_name", type="string", online_type="varchar(100)"),
             Feature("last_name", type="string", online_type="varchar(100)"),
             Feature("phone", type="string", online_type="varchar(50)"),
-            Feature("line_count", type="int"),
+            Feature("line_count", type="bigint"),
             _json_feature("purchases"),
             Feature("migrated_at", type="timestamp"),
         ],
